// hybris/bin/modules/smartedit/smartedit/web/app/common/services/interfaces/IExperience.ts
/*
 * Copyright (c) 2019 SAP SE or an SAP affiliate company. All rights reserved.
 */
import {Payload, Primitive, TypedMap} from 'smarteditcommons/dtos';
import {ISite} from './ISite';
import {ILanguage} from "../language/LanguageService";

/**
 * @ngdoc interface
 * @name smarteditServicesModule.interface:IExperience
 * @description
 * IExperience - Interface for experience information
 */
export interface IExperience extends Payload {
	catalogDescriptor: IExperienceCatalogDescriptor;
	siteDescriptor: ISite;
	productCatalogVersions: IExperienceCatalogVersion[];
	time: string;

	languageDescriptor?: ILanguage;
	pageId?: string;
	pageContext?: IExperiencePageContext;
	[index: string]: Primitive | Primitive[] | Payload | Payload[];
}

/**
 * @ngdoc interface
 * @name smarteditServicesModule.interface:IExperienceCatalogDescriptor
 * @description
 * IExperienceCatalogDescriptor interface acts as a pointer to catalog and catalogVersion of the experience
 */
export interface IExperienceCatalogDescriptor extends Payload {
	active: boolean;
	catalogId: string;
	catalogVersion: string;
	catalogVersionUuid: string;
	name: TypedMap<string>;
	siteId: string;
}

export interface IExperienceCatalogVersion extends Payload {
	active: boolean;
	catalog: string;
	catalogName: TypedMap<string>;
	catalogVersion: string;
	uuid: string;
}

export interface IExperiencePageContext extends Payload {
	catalogId: string;
	catalogName: TypedMap<string>;
	catalogVersion: string;
	catalogVersionUuid: string;
	siteId: string;
	active: boolean;
}

export interface IDefaultExperienceParams {
	siteId: string;
	catalogId: string;
	catalogVersion: string;
	pageId?: string;
}

export interface IExperienceParams extends IDefaultExperienceParams {
	time?: string;
	productCatalogVersions?: string[];
	language?: string;
}